Add option to rename scaffolds
A new option was added to rename scaffolds to "1" and make the positions cumulative (concatenating all scaffolds). This is useful for datasets based on reference genomes that are not chromosome level assemblies. The script now also assumes a constant recombination rate of 2 cM/Mb and uses the SNP positions to compute the Morgan positions.